requirements.txt
grpcio-tools
flask
Jinja2
pytest